+++ src/memory.h
@@ -25,24 +25,24 @@ namespace Lighthouse {
     class Memory : public QObject
     {
         Q_OBJECT
-        Q_PROPERTY(int free READ getFree NOTIFY freeChanged)
-        Q_PROPERTY(int total READ getTotal NOTIFY totalChanged)
+        Q_PROPERTY(unsigned long long free READ getFree NOTIFY freeChanged)
+        Q_PROPERTY(unsigned long long total READ getTotal NOTIFY totalChanged)
         Q_PROPERTY(int summaryValue READ getSummaryValue NOTIFY summaryValueChanged)
     public:
         explicit Memory(QObject *parent = 0);
 
-        int getFree() const;
-        int getTotal() const;
+        unsigned long long getFree() const;
+        unsigned long long getTotal() const;
         int getSummaryValue() const;
     private:
-        int fFree;
-        int fTotal;
+        unsigned long long fFree;
+        unsigned long long fTotal;
     signals:
         void freeChanged(int free);
         void totalChanged(int total);
         void summaryValueChanged(int value);
     public slots:
-        void setMemory(int total, int free);
+        void setMemory(unsigned long long total, unsigned long long free);
     };
 
 }
